Extract audit into separate mixin, rename actions button component, code cleanup
