+++ frontend/src/pages/team/Applications/components/ApplicationDevice.vue
@@ -22,7 +22,6 @@
     <div class="flex justify-end text-sm">
         <EditorLink
             :instance="device"
-            :url="device.editor?.url"
             :editorDisabled="false"
             :disabled="!device.editor?.enabled || !device.editor?.connected || !device.editor?.local"
             disabledReason="Device must be running, in developer mode and have the editor enabled and connected"
@@ -32,7 +31,7 @@
 
 <script>
 import StatusBadge from '../../../../components/StatusBadge.vue'
-import AuditEventsService from '../../../../services/audit-events.js'
+import AuditMixin from '../../../../mixins/Audit.js'
 import DaysSince from '../../../application/Snapshots/components/cells/DaysSince.vue'
 import DeviceModeBadge from '../../../device/components/DeviceModeBadge.vue'
 import EditorLink from '../../../instance/components/EditorLink.vue'
@@ -45,28 +44,11 @@ export default {
         DaysSince,
         StatusBadge
     },
+    mixins: [AuditMixin],
     props: {
         device: {
             type: Object,
             required: true
-        }
-    },
-    setup () {
-        return {
-            AuditEvents: AuditEventsService.get()
-        }
-    },
-    methods: {
-        labelForAuditLogEntry (eventName) {
-            if (!eventName) return 'Unknown Event'
-            if (this.AuditEvents[eventName]) {
-                return this.AuditEvents[eventName]
-            }
-            let labelText = eventName
-            labelText = labelText.replace(/[-._:]/g, ' ')
-            labelText = labelText.replace(/\b\w/g, l => l.toUpperCase())
-            labelText = labelText.replace(/([a-z])([A-Z])/g, '$1 $2')
-            return labelText
         }
     }
 }
